+++ api/models/schema/user.js
@@ -76,14 +76,14 @@ UserSchema.plugin(autopopulate);
 UserSchema.pre('save', true, function (next, done) {
     next(); // in parallel ^
 
-    if (this.isModified('picture') || this.isNew) {
+    if (this.isModified('picture') || (this.isNew && this.picture)) {
         if (this._oldPicture) {
             // delete old one
             fs.unlink(`${upload.dest}${this._oldPicture}`);
         }
 
         // copy photo from tmp
-        // todo save as Buffer in the database and delete
+        // todo save as Buffer in the database and delete ?
         fs.rename(
             `${upload.destTmp}${this.picture}`,
             `${upload.dest}${this.picture}`,
